feat(energies) create endpoint with new schema

+++ src/modules/users/users.controller.ts
@@ -11,7 +11,7 @@ export const usersController = {
 
   show: async (req: Request, res: Response) => {
     const { query, params } = res.locals.validatedData;
-
+    if (!params) throw new Error('ID harus diisi');
     const userId = params?.id ? Number(params.id) : undefined;
 
     const data = await usersService.show(userId, query);
@@ -19,15 +19,18 @@ export const usersController = {
   },
 
   patch: async (req: Request, res: Response) => {
-    // 3. Ambil id dari validatedData saja biar konsisten dengan Zod
     const { params, body } = res.locals.validatedData;
+
+    if (!params) throw new Error('ID harus diisi');
+
     const data = await usersService.patch(Number(params.id), body);
     return res200({ res, message: 'Berhasil Mengubah Data User', data });
   },
 
   remove: async (req: Request, res: Response) => {
-    // 4. Sama seperti patch, ambil dari validatedData.params
     const { params } = res.locals.validatedData;
+    if (!params) throw new Error('ID harus diisi');
+
     const data = await usersService.remove(Number(params.id));
     return res200({ res, message: 'Berhasil Menghapus Data User', data });
   },
